% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Alpha0}
\alias{Alpha0}
\title{Estimate \eqn{\alpha} under \eqn{H_{0}}.}
\usage{
Alpha0(X, Ri, y)
}
\arguments{
\item{X}{Design matrix for alpha, numeric.}

\item{Ri}{Inverse correlation structure, numeric.}

\item{y}{Response, numeric.}
}
\description{
Estimates coefficient on nuisance parameter under the \eqn{H_{0}:\beta=0}.
}
