% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{coef.fit}
\alias{coef.fit}
\title{Extract Coefficients from a Fitted Regression Model}
\usage{
\method{coef}{fit}(object, ...)
}
\arguments{
\item{object}{A \code{fit} object.}

\item{...}{Unused.}
}
\description{
Returns the estimated regression coefficients from an object of class \code{fit}.
}
