% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combinedlist.R
\name{combined.listDirect}
\alias{combined.listDirect}
\title{Combined List Estimator}
\usage{
combined.listDirect(formula, data = parent.frame(), treat = "treat",
  direct = "direct", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. Should be of the form Y ~ T + X1 + X2, where Y is the list response, T is the treatment indicator, and X1, X2, etc are pretreatment covariates. It is recommended that T be a numeric variable whose values are 0 for subjects in control and 1 for subjects in treatment.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which combined.list is called. It is good practice to include all variables used in the estimation (list response, treatment indicator, direct response, and optional pre-treatment covariates) in a dataframe rather than calling data from the global environent.}

\item{treat}{a character string giving the name of the treatment variable. Defaults to "treat".}

\item{direct}{a character string giving the name of the direct response variable. Defaults to "direct". The direct response variable itself must only contain the values 0 and 1, where 1 refers to subjects who answered "Yes" to the direct question.}
}
\value{
a list containing conventional, direct, and combined prevalence estimates with associated standard errors as well as the results of two placebo tests.
}
\description{
This function implements the combined list estimator described in Aronow, Coppock, Crawford, and Green (2015): Combining List Experiment and Direct Question Estimates of Sensitive Behavior Prevalence
}
\examples{
# Define subject types.

# Truthfully respond "Yes" to direct question
N.trueadmitter <- 500

# Falsely respond "No" to direct question
N.withholder <- 500

# Truthfully respond "No" to direct question
N.innocent <- 500

type <- rep(c("TA", "WH", "IN"), times=c(N.trueadmitter, N.withholder, N.innocent))
N <- length(type)

# Generate a predictive pre-treatment covariate "X")
X <- rnorm(N, sd = 2)

# Generate list response potential outcomes

# Control potential outcome is related to "X"
Y0 <- as.numeric(cut(X + runif(N), breaks = 4))

# Treated potential outcome is 1 higher for true admitters and withholders
Y1 <- Y0 + ifelse(type \%in\% c("TA", "WH"), 1, 0)

# Conduct random assignment
Z <- rbinom(N, 1, 0.5)

# Reveal direct question responses
D <- ifelse(type=="TA", 1, 0)

# Reveal list responses
Y <- Z*Y1 + (1-Z)*Y0

# Wrap into dataframe
df <- data.frame(Y, Z, D, X)

# Conduct estimation without covariate adjustment
out.1 <- combined.listDirect(formula = Y ~ Z, data = df, treat = "Z", direct = "D")
summary(out.1)

# Conduct estimation with covariate adjustment
out.2 <- combined.listDirect(formula = Y ~ Z + X, data = df, treat = "Z", direct = "D")
summary(out.2)
}

