% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedListDirect.R, R/list-package.R
\docType{data}
\name{combinedListExps}
\alias{combinedListExps}
\title{Five List Experiments with Direct Questions}
\format{
A data frame with 1023 observations and 23 variables

A data frame with 1023 observations and 23 variables
}
\usage{
data(combinedListExps)
}
\description{
A dataset containing the five list experiments in 
Aronow, Coppock, Crawford, and Green (2015)

A dataset containing the five list experiments in Aronow, Coppock, Crawford,
and Green (2015)
}
\keyword{datasets}
