% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{D}
\alias{D}
\title{Simulated Data}
\format{A list of vectors and matrices. 
\describe{ 
   \item{y}{Numeric outcome vector.}
   \item{Z}{Numeric model matrix.}
   \item{Beta}{Regression coefficients used to generate the outcome.}
}}
\usage{
D
}
\description{
Example data for 1000 subjects. Outcomes were generated by adding standard
normal residuals to the vector of means. The outcome means depend on a model
matrix containing an intercept and three N(0,1) covariates.
}
\keyword{datasets}
