% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exchCorr}
\alias{exchCorr}
\title{Generate Exchangable Correlation Structure}
\usage{
exchCorr(n, r)
}
\arguments{
\item{n}{Dimension}

\item{r}{Correlation}
}
\description{
Generate Exchangable Correlation Structure
}
