% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Fit.LinReg}
\alias{Fit.LinReg}
\title{Fit Linear Regression Model}
\usage{
Fit.LinReg(y, X = NULL, w = NULL, sig = 0.05)
}
\arguments{
\item{y}{Numeric outcome vector.}

\item{X}{Numeric model matrix. Include an intercept.}

\item{w}{Optional vector of known, observation-specific weights.}

\item{sig}{Significance level, for CIs.}
}
\description{
Fit Linear Regression Model
}
\examples{
\dontrun{
# Generate data
set.seed(100);
# Design
X = cbind(1,matrix(rnorm(4e3),nrow=1e3));
colnames(X) = c("int",paste0("x",seq(1:4)));
b = c(1,-0.1,0.2,-0.1,0);
# Linear predictor
h = as.numeric(MMP(X,b));
# Outcome
y = h+rnorm(1e3);
# Model fit
M = Fit.LinReg(y=y,X=X);
show(M);
}
}
