% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitNorm}
\alias{fitNorm}
\title{Normal Model}
\usage{
fitNorm(y, Z, estT, t)
}
\arguments{
\item{y}{Outcome.}

\item{Z}{Model matrix.}

\item{estT}{Logical indicating tau should be estimated. If false, provide the
known value.}

\item{t}{Variance component, if known.}
}
\description{
Normal Model
}
