% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ict.hausman.test.R
\name{ict.hausman.test}
\alias{ict.hausman.test}
\title{Hausman Specification Test for Two List Experiment Regression Fit Objects}
\usage{
ict.hausman.test(fit1, fit2)
}
\arguments{
\item{fit1}{}

\item{fit2}{}
}
\description{
Hausman Specification Test for Two List Experiment Regression Fit Objects
}
