% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{kNormScore}
\alias{kNormScore}
\title{Kernelized Score Test for Linear Models}
\usage{
kNormScore(G, W, X, y, R)
}
\arguments{
\item{G}{Matrix whose regression coefficient is zero in the null model.}

\item{W}{Weight matrix.}

\item{X}{Matrix whose regression coefficient is unconstrained in the null 
model.}

\item{y}{Response vector.}

\item{R}{Optional fixed correlation structure for variance component. Default
is identity.}
}
\value{
List including the \code{Score}, the \code{Eigenvalues} of the
  mixutre distribution, and the \code{p} value.
}
\description{
Performs the score test of the null hypothesis \eqn{\beta = 0} for a normal 
linear model of the form: \deqn{y = G\beta + X\alpha + \epsilon} Here 
\eqn{\epsilon} is distributed as \eqn{N(0,\tau R)}, where \eqn{\tau} is a 
variance component, and \eqn{R} is an \eqn{n x n} correlation structure. The 
kernelized score statistic takes the form \deqn{T_{K} =
\tilde{\epsilon}'K\tilde{\epsilon}} Here \eqn{\tilde{\epsilon}} is the
residual estimated under \eqn{H_{0}}, and \eqn{K = WG(WG)'} is the kernel matrix.
}
\examples{
# Genotypes
G = NormalScoreTest::G[,1:10];
maf = apply(G,MARGIN=2,FUN=mean)/2;
Gs = scale(G);
# Weights
W = round(diag((dbeta(x=maf,1,25))),digits=6);
# Phenotype
y = NormalScoreTest::Y[,1];
Results = NormalScoreTest::kNormScore(G=Gs,W=W,X=NormalScoreTest::X,y=y);
}
