% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knScore}
\alias{knScore}
\title{Kernalized Score Statistic for Normal Linear Model}
\usage{
knScore(X, L, Ri, y, tau)
}
\arguments{
\item{X}{Matrix whose regression coefficient is unconstrained in the null
model.}

\item{L}{Cholesky decomposition of kernel matrix.}

\item{Ri}{Inverse correlation structure, numeric.}

\item{y}{Response, numeric.}

\item{tau}{Estimate of tau}
}
\description{
Calculates a kernelized score statistic of \eqn{H_{0}:\beta = 0} in the form
\deqn{T_{K} = \epsilon'K\epsilon} where \eqn{K = LL'}
}
