% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\docType{package}
\name{LiST-help}
\alias{LiST-help}
\alias{LiST-help-package}
\title{LiST: Linear Score Tests}
\description{
Implementation of score tests for linear models. \code{\link{Score}} performs a basic
score test, providing either an asymptotic or perturbation p-value. \code{\link{wScore}}
performs a weighted score test. \code{\link{rScore}} performs repeated association tests.
\code{\link{pScore}} performs score tests on a model previously fit using \code{\link{Fit.LinReg}}.
}
\author{
Zachary R. McCaw
}
