% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{lmScoreTest}
\alias{lmScoreTest}
\title{Linear Model Score Test}
\usage{
lmScoreTest(G, X, y, R)
}
\arguments{
\item{G}{Matrix whose regression coefficient is zero in the null model.}

\item{X}{Matrix whose regression coefficient is unconstrained in the null
model.}

\item{y}{Response vector.}

\item{R}{Optional fixed correlation structure for variance component. Default
is identity.}
}
\value{
Vector include the score statistic, the degrees of freedom, and a
  p-value based on the chi-square distribution.
}
\description{
Performs the score test of the null hypothesis \eqn{\beta = 0} for a normal 
linear model of the form: \deqn{y = G\beta + X\alpha + \epsilon} Here
\eqn{\epsilon} is distributed as \eqn{N(0,\tau R)}, where \eqn{\tau} is a 
variance component, and \eqn{R} is an \eqn{n x n} correlation structure.
}
\examples{

}
