% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-package.R
\docType{data}
\name{mis}
\alias{mis}
\title{The 1994 Multi-Investigator Survey}
\format{
A data frame containing the following 6 variables for 1171
observations.

\tabular{llll}{ y \tab numeric \tab the number of items that make
respondents angry \tab 0 - 4 \cr sensitive \tab numeric \tab whether or not
the sensitive item (asked directly) makes respondents angry \tab 0 - 1 \cr
south \tab numeric \tab whether or not a respondents live in a southern
state \tab 0 - 1 \cr male \tab numeric \tab whether or not a respondent is
male \tab 0 - 1 \cr college \tab numeric \tab whether or not a respondent
attended some college \tab 0 - 1 \cr age \tab numeric \tab age of a
respondent divided by 10 \cr democrat \tab numeric \tab whether not a
respondent identifies as a Democrat \tab 0 - 1 \cr republican \tab numeric
\tab whether not a respondent identifies as a Republican \tab 0 - 1\cr
independent \tab numeric \tab whether not a respondent identifies as an
independent \tab 0 - 1 \cr treat \tab numeric \tab treatment status \tab 0 -
1 \cr list.data \tab numeric \tab indicator for list experiment subset
(treatment and control groups) \tab 0 - 1 \cr sens.data \tab numeric \tab
indicator for direct sensitive item subset \tab 0 - 1 \cr

}
}
\source{
The full data set is available at SDA (Survey Documentation and
Analysis; \url{http://sda.berkeley.edu/D3/Multi/Doc/mult.htm})
}
\description{
This dataset is a subset of the 1994 Multi-Investigator Survey and contains
the item count technique or the list experiment.  The main question reads as
follows: \verb{ Now I'm going to read you four things that sometimes make
people angry or upset.  After I read all (three/four), just tell me HOW MANY
of them upset you.  (I don't want to know which ones, just how many.)  (1)
"the federal government increasing the tax on gasoline;" (2) "professional
athletes getting million-dollar-plus salaries;" (3) "requiring seatbelts be
used when driving;" (4) "large corporations polluting the environment;" (5)
"black leaders asking the government for affirmative action."}
}
\details{
where the last item is presented only with the treatment group and the
control list only contains the first three items.

The survey also includes a question in which attitudes toward the sensitive
item are asked directly.

\verb{ Now I'm going to ask you about another thing that sometimes makes
people angry or upset. Do you get angry or upset when black leaders ask the
government for affirmative action?}
}
\keyword{datasets}
