% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misreg.R
\name{misreg}
\alias{misreg}
\title{Regression Models for Misreporting}
\usage{
misreg(data, rhs = NULL, y, dir, treat, J, sensitive.treatment = TRUE,
  method = "ratio", model.extremes = FALSE, two.way = FALSE,
  bayes = FALSE, spec.test = FALSE, infer.design = TRUE, design,
  nls.options, ml.options)
}
\arguments{
\item{data}{data frame containing all variables}

\item{rhs}{formula with predictor variables}

\item{y}{vector of indirect responses}

\item{dir}{vector of direct responses. should be coded such that 1 implies that the respondent has the sensitive trait.  can contain missing values.}

\item{treat}{name of list experiment treatment indicator in data}

\item{J}{number of innocuous control itmes}

\item{sensitive.treatment}{indicates whether the treatment item is sensitive (I have the sensitive trait) or not (I do not have the sensitive trait). TRUE/FALSE}

\item{method}{one of lm, nls, ml, or ratio}

\item{model.extremes}{TRUE/FALSE}

\item{two.way}{TRUE/FALSE}

\item{bayes}{indicates whether to regularize logistic regression coefficients using the method of Gelman et al.  Note that continuous coefficients should be standardized by two standard deviations.  TRUE/FALSE, default is FALSE}

\item{spec.test}{default is TRUE if method = "ml", otherwise FALSE}

\item{infer.design, }{default is TRUE}

\item{design}{an N by 2 matrix that indicates if the respondent is assigned to receive the direct question (column 1) and the list experiment (column 2)}

\item{nls.options}{a list of options that specifies whether to iterate the fit or to estimate in one-step}

\item{ml.options}{can specifiy tolerance and maximum number of iterations}
}
\value{
\code{misreg} returns an object of class "misreg".
}
\description{
Note: This function is in development.  The ratio, NLS, and ML estimators 
for list experiments with the standard wording (i.e., the treatment item 
directly gauges the sensitive item prevalence) without floor effects are 
implemented.
}
\details{
This function implements the methods introduced in Chou (N.d.) and 
Eady (2017) for estimating multiple regression models of misreporting. 
The methods can be applied to survey designs that combine direct 
questioning with the list experiment, a popular tool for eliciting 
sensitive attitudes (see Blair and Imai 2011).
}
\references{
Chou, Winston. (2018) ``Lying on Surveys.'' Technical report, Princeton University.
}
\author{
Winston Chou, Princeton University, \email{wchou@princeton.edu}
}
