% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nScore}
\alias{nScore}
\title{Score Statistic for Normal Linear Model}
\usage{
nScore(X, G, Ri, y, tau)
}
\arguments{
\item{X}{Matrix whose regression coefficient is unconstrained in the null
model.}

\item{G}{Matrix whose regression coefficient is zero in the null model.}

\item{Ri}{Inverse correlation structure, numeric.}

\item{y}{Response, numeric.}

\item{tau}{Estimate of tau}
}
\description{
Calculates the score statistic of \eqn{H_{0}:\beta_{G}=0} in the form
\deqn{T_{S}=y'QG(G'QG)^{-1}G'Qy}
}
