% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{predict.fit}
\alias{predict.fit}
\title{Prediction from a Fitted Regression Model}
\usage{
\method{predict}{fit}(object, ..., X = NULL)
}
\arguments{
\item{object}{A \code{fit} object.}

\item{...}{Unused.}

\item{X}{Numeric model matrix. If omitted, returns predictions for the 
observations used to fit the model}
}
\value{
A numeric vector of predictions.
}
\description{
Returns the estimated residuals from an object of class \code{fit}.
}
