% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pScore.R
\name{pScore}
\alias{pScore}
\title{Prefit Model Score Test}
\usage{
pScore(M0, G, X, method = "asymptotic", B = 1000, parallel = F)
}
\arguments{
\item{M0}{Prefit null model.}

\item{G}{Covariates tested jointly for association with the outcome.}

\item{X}{Model matrix used to fit \code{M0}.}

\item{method}{Either "asymptotic" or "perturbation".}

\item{B}{Score perturbations.}

\item{parallel}{Run in parallel? Must reigster parallel backend first.}
}
\description{
Performs a joint score test on the covariates in G using a previously fit null model \code{M0}.
}
\examples{
\dontrun{
Fit null model
M0 = Fit.LinReg(y=y,X=X);
# Matrix to test for association
G = matrix(rnorm(2e3),nrow=1e3);
# Prefit model score test
pScore(M0=M,G=G,X=X);
# Standard score test
Score(y=y,X=cbind(X,G),L=c(rep(F,ncol(X)),rep(T,ncol(G))));
}
}
