% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residV}
\alias{residV}
\title{Residual Variance}
\usage{
residV(X, R, tau)
}
\arguments{
\item{X}{Desigm matrix for alpha, numeric.}

\item{R}{Correlation structure, numeric.}

\item{tau}{Variance component, numeric.}
}
\description{
Calculate the variance of \eqn{y-X\tilde{\alpha}} under \eqn{H_{0}}.
}
