% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rScore.R
\name{rScore}
\alias{rScore}
\title{Repeated Score Test}
\usage{
rScore(y, G, X, parallel = F)
}
\arguments{
\item{y}{Numeric response vector.}

\item{G}{Numeric matrix of covariates for the target outcome whose regression
coefficients are zero under the null.}

\item{X}{Numeric model matrix included in the null model.}

\item{parallel}{Run association tests if parallel? Must register parallel 
backend first.}
}
\value{
Vector of asymptotic p-values, one for each column of \code{G}.
}
\description{
Individually tests each column of \code{G} for association with \code{y} 
while adjusting for \code{X}. \code{G} may contain missing elements, although
the remaining model matrices may not. \code{rScore} accelerates association
testing by recycling the same null model for each hypothesis test.
}
\examples{
\dontrun{
# Genotypes
G = replicate(rbinom(n=1e3,size=2,prob=0.25),n=2000);
G[sample(length(G),size=floor(0.01*length(G)),replace=F)] = NA;
storage.mode(G) = "numeric";
doMC::registerDoMC(cores=2);
# Repeated Score Test
P = rScore(y=y,G=G,X=X,parallel=T);
}
}
