% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SchurC}
\alias{SchurC}
\title{Schur Complement}
\usage{
SchurC(Igg, Ihh, Igh, inv)
}
\arguments{
\item{Igg}{First information matrix.}

\item{Ihh}{Second information matrix.}

\item{Igh}{Cross information matrix.}

\item{inv}{Return the inverse of}
}
\description{
Calcualtes the matrix \deqn{I_{gg}-I_{gh}I_{hh}^{-1}I_{gh}'}
}
