% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{Score.nlm}
\alias{Score.nlm}
\title{Score Test for Normal Linear Models}
\usage{
Score.nlm(y, Z, L, b10)
}
\arguments{
\item{y}{Numeric response vector.}

\item{Z}{Numeric model matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under 
the null. Defaults to zero.}
}
\value{
A numeric vector containing the score statistic, the degrees of
  freedom, and a p-value estimated using the chi-square distribution.
}
\description{
Performs a score test of \eqn{H_{0}:\beta_{1}=\beta_{10}}. The score test is 
specified using a logical vector \code{L}, with as many entries as columns in
the model matrix \code{Z}. The values of \code{L} set to \code{T} are fixed
at \eqn{\beta_{10}} under the null. The values of \code{L} set to \code{F}
are estimated.
}
\examples{
y = D$y;
Z = D$Z;
# Overall test for covariate effects, should reject.
Score.nlm(y=y,Z=Z,L=c(FALSE,TRUE,TRUE,TRUE));
# Individual test of b3=0, should not reject.
Score.nlm(y=y,Z=Z,L=c(FALSE,FALSE,FALSE,TRUE));
# Joint test for the effects of b1=0 and b2=0, should reject.
Score.nlm(y=y,Z=Z,L=c(FALSE,TRUE,TRUE,FALSE));
# Individual test of b1=0.1, should not reject.
Score.nlm(y=y,Z=Z,L=c(FALSE,TRUE,FALSE,FALSE),b10=c(0.1));
}
