% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{Score}
\alias{Score}
\title{Standard Score Test}
\usage{
Score(y, X, L, b10 = NULL, method = "asymptotic", B = 1000,
  parallel = F)
}
\arguments{
\item{y}{Numeric response vector.}

\item{X}{Numeric model matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{method}{Either "asymptotic" or "perturbation".}

\item{B}{Score perturbations.}

\item{parallel}{Run in parallel? Must reigster parallel backend first.}
}
\value{
A numeric vector containing the score statistic, the degrees of 
  freedom, and a p-value.
}
\description{
Tests the hypothesis that a subset of the regression coefficients are fixed 
at a reference value. Specifically, let \eqn{\beta} denote the regression 
coefficient. Partition \eqn{\beta=(\beta_{1},\beta_{2})}. Suppose that 
interest lies in testing that \eqn{\beta_{1}} is fixed at \eqn{\beta_{10}}. 
\code{Score} performs a score test of \eqn{H_{0}:\beta_{1}=\beta_{10}}. The
test is specified using a logical vector \code{L}, with as many entries as
columns in the model matrix \code{X}. The values of \code{L} set to \code{T}
are constrained under the null, while values of \code{L} set to \code{F} are
estimated under the null.
}
\examples{
\dontrun{
# See Vignette for Data Generation.
# Hypothesis test
L = c(F,F,F,F,T);
# Asymptotic p-value
Score(y=y,X=X,L=L,method="asymptotic");
# Perturbation p-value
Score(y=y,X=X,L=L,method="perturbation");
}
}
