% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScoreB}
\alias{ScoreB}
\title{Calculate Score under H0}
\usage{
ScoreB(e0, G, R, tau)
}
\arguments{
\item{e0}{Residuals under H0, numeric.}

\item{G}{Design matrix for beta, numeric.}

\item{R}{Correlation structure, numeric.}

\item{tau}{Variance component, numeric.}
}
\description{
Calculate the score vector for \eqn{\beta} under \eqn{H_{0}:\beta=0}.
}
