% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{sScoreTest}
\alias{sScoreTest}
\title{Serial Score Test}
\usage{
sScoreTest(G, X, y, R, parallel = F)
}
\arguments{
\item{G}{Matrix whose regression coefficients are zero under the null model.}

\item{X}{Matrix whose regression coefficient is unconstrained in the null 
model.}

\item{y}{Response vector.}

\item{R}{Optional fixed correlation structure for variance component. Default
is identity.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
Matrix with one row per column of G. Each row includes the score
  statistic, the degrees of freedom, and a p-value based on the chi-square
  distribution.
}
\description{
For each column of \eqn{G_{j}}, tests \eqn{H_{0}:\beta_{j}=0} in the model: 
\deqn{y = G_{j}\beta_{j}+X\alpha+\epsilon} See \code{\link{ScoreTest}} for 
model description.
}
\examples{
# Genotype matrix
G = ScoreTest::G;
# Standardizing genotype matrix
Gs = scale(t(G));
# Apply score test to each column of Gs
Results = ScoreTest::sScoreTest(G=Gs,X=ScoreTest::X,y=ScoreTest::Y);
}
