% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tau0}
\alias{Tau0}
\title{Estimate \eqn{tau} under \eqn{H_{0}}.}
\usage{
Tau0(X, Ri, y, a)
}
\arguments{
\item{X}{Overall design matrix, numeric.}

\item{Ri}{Inverse correlation structure, numeric.}

\item{y}{Response, numeric.}

\item{a}{Estimated regression coefficient}
}
\description{
Estimates the variance component \eqn{\tau} under the reduced
model where \eqn{\beta = 0}.
}
