% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tau1}
\alias{Tau1}
\title{Estimate \eqn{tau} under \eqn{H_{1}}.}
\usage{
Tau1(D, Ri, y)
}
\arguments{
\item{D}{Overall design matrix, numeric.}

\item{Ri}{Inverse correlation structure, numeric.}

\item{y}{Response, numeric.}
}
\description{
Estimates the variance component \eqn{\tau} under the reduced
model where \eqn{\beta = 0}.
}
