% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{vcov.fit}
\alias{vcov.fit}
\title{Extract Covariance Matrix from a Fitted Regression Model}
\usage{
\method{vcov}{fit}(object, ..., inv = F)
}
\arguments{
\item{object}{A \code{fit} object.}

\item{...}{Unused.}

\item{inv}{Invert information matrix?}
}
\value{
A numeric matrix.
}
\description{
Returns the (inverse) Fisher information of the regression
coefficient \eqn{\hat{\beta}}.
}
