% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wNormScore}
\alias{wNormScore}
\title{Weighted Normal Score Statistic}
\usage{
wNormScore(y, X1, b, X2, estT, t, useK, K, W)
}
\arguments{
\item{y}{Outcome.}

\item{X1}{Model matrix whose coefficient is fixed.}

\item{b}{Values at which the coefficients of X1 are fixed.}

\item{X2}{Model matrix whose coefficient is estimated.}

\item{estT}{Logical indicating tau should be estimated. If false, provide the
known value.}

\item{t}{Variance component, if known.}

\item{useK}{Logical indicating that a known covariance structure is supplied.}

\item{K}{Covariance structure, if known.}

\item{W}{Weight matrix}
}
\description{
Weighted Normal Score Statistic
}
