% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{wScore.nlm}
\alias{wScore.nlm}
\title{Weighted Score Test for Normal Models}
\usage{
wScore.nlm(y, Z, W, L, b0, tau, K)
}
\arguments{
\item{y}{Response vector.}

\item{Z}{Numeric model matrix.}

\item{W}{Numeric weight matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{b0}{Value of the regression coefficient for the selected columns under 
the null. Defaults to zero.}

\item{tau}{Value of the scale parameter if known. If omitted, \eqn{\tau} is 
estimated.}

\item{K}{Optional fixed correlation structure for variance component. Default
is identity.}
}
\value{
Vector including the score statistic, the degrees of freedom, and a 
  p-value based on the chi squared distribution.
}
\description{
Performs a weighted score test of \eqn{H_{0}:\beta_{1}=\beta_{10}} using
the test statistic \deqn{T_{W}=S'WWS} Here \eqn{S} is the score vector and
\eqn{W} is a summetric weight matrix.
}
