% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScore.R
\name{wScore}
\alias{wScore}
\title{Weighted Score Test}
\usage{
wScore(y, X, L, b10 = NULL, w = NULL, method = "asymptotic",
  B = 1000, parallel = F)
}
\arguments{
\item{y}{Numeric response vector.}

\item{X}{Numeric model matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{w}{Weights for the selected regression coefficients under the null.
Defaults to the one vector.}

\item{method}{Either "asymptotic" or "perturbation".}

\item{B}{Score perturbations.}

\item{parallel}{Run in parallel? Must reigster parallel backend first.}
}
\value{
A numeric vector containing the score statistic, the degrees of 
  freedom, and a p-value.
}
\description{
Tests the hypothesis that a subset of the regression coefficients are fixed 
at a reference value, affording different weights to different coefficients. 
The weights are renormalized internally such that the observed test statistic
is one.
}
\examples{
\dontrun{
# See Vignette for Data Generation.
# Hypothesis test
L = c(F,F,F,T,T);
# Asymptotic p-value
wScore(y=y,X=X,L=L,w=c(2,1),method="asymptotic");
# Perturbation p-value
wScore(y=y,X=X,L=L,w=c(2,1),method="perturbation");
}
}
