% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated Covariates}
\format{A numeric matrix with 1000 rows and 4 columns
\describe{
   \item{Age}{Age.}
   \item{Sex}{Sex.}
   \item{PC1}{First principal component.}
   \item{PC2}{Second principal component.}
}}
\usage{
X
}
\description{
Age, sex, and two principal components of the centered and scaled subject by
locus genotype matrix. Age was drawn from a gamma distribution with mean 50
and variance 10. Sex was drawn from a Bernoulli distribuiton with expectation
1/2.
}
\keyword{datasets}
