% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated Phenotype}
\format{A numeric matrix with 1000 rows and 1 columns
\describe{
   \item{YN}{Normal phenotype.}
}}
\usage{
Y
}
\description{
Phenotypes simulated under the null hypothesis of no genotypic effect. A
subject specific linear predictor was calculated based on the covaraites. For
\code{Y.ind}, the residuals were drawn from a N(0,I) distribution. For \code{Y.corr},
the residuals were drawn from a N(0,R) distribution
}
\keyword{datasets}
