% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Alpha0}
\alias{Alpha0}
\title{Estimate Alpha under H0}
\usage{
Alpha0(X, R, y)
}
\arguments{
\item{X}{Design matrix for alpha, numeric.}

\item{R}{Correlation structure, numeric.}

\item{y}{Response, numeric.}
}
\description{
Estimates coefficient on nuisance parameter under the \eqn{H_{0}:\beta=0}.
}
