% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Info}
\alias{Info}
\title{Estimate Information Matrix}
\usage{
Info(X, G, R, tau)
}
\arguments{
\item{X}{Design matrix for alpha, numeric.}

\item{G}{Design matrix for beta, numeric.}

\item{R}{Correlation structure, numeric.}

\item{tau}{Variance component, numeric.}
}
\description{
Estimate information matrix for \eqn{(\beta,\alpha)}.
}
