% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rScoreTest.R
\name{rScore.nlm}
\alias{rScore.nlm}
\title{Repeated Score Test for Normal Models}
\usage{
rScore.nlm(y, X1, X2, tau, K)
}
\arguments{
\item{y}{Response vector.}

\item{X1}{Numeric matrix of covariates whose regression coefficients are 
constrained under the null.}

\item{X2}{Numeric matrix of covariates whose regression coefficients are 
unconstrained under the null.}

\item{tau}{Vaule of the scale parameter if known. If omitted, \eqn{\tau} is 
estimated.}

\item{K}{Optional fixed correlation structure for variance component. Default
is identity.}
}
\value{
Vector of p-values, one for each column of \code{X1}, based on the 
  chi square distribution.
}
\description{
Individually tests \eqn{H_{0}:\beta_{j}=0} for each column of \code{X1}, 
adjusting for \code{X2}.
}
