% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{Score.nlm}
\alias{Score.nlm}
\title{Score Test for Normal Models}
\usage{
Score.nlm(y, Z, L, b0, K, tau)
}
\arguments{
\item{y}{Response vector.}

\item{Z}{Numeric model matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{b0}{Value of the regression coefficient for the selected columns under 
the null. Defaults to zero.}

\item{K}{Optional fixed correlation structure for variance component. Default
is identity.}

\item{tau}{Vaule of the scale parameter if known. If omitted, \eqn{\tau} is 
estimated.}
}
\value{
Vector including the score statistic, the degrees of freedom, and a 
  p-value based on the chi squared distribution.
}
\description{
Performs a score test of \eqn{H_{0}:\beta_{1}=\beta_{10}}.
}
