% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tau1}
\alias{Tau1}
\title{Estimates Tau under H1}
\usage{
Tau1(D, R, y)
}
\arguments{
\item{D}{Overall design matrix, numeric.}

\item{R}{Correlation structure, numeric.}

\item{y}{Response, numeric.}
}
\description{
Estimate the variance component \eqn{\tau} under the full model.
}
