% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScoreTest.R
\name{wScore.nlm}
\alias{wScore.nlm}
\title{Variance Component Score Test}
\usage{
wScore.nlm(y, Z, L, w)
}
\arguments{
\item{y}{Response vector.}

\item{Z}{Numeric model matrix.}

\item{L}{Logical vector, with as many entires as columns in the model matrix,
indicating which columns have fixed coefficients under the null.}

\item{w}{Numeric weight vector.}
}
\value{
Returns a list with the following components:
\item{Score}{The weighted score statistic.}
\item{Lambda}{The eigenvalues of \eqn{\Xi}.}
\item{p}{The estimated p-value.}
}
\description{
Performs a weighted score test of \eqn{H_{0}:\beta_{1}=0}. The score
test is specified using a logical vector \code{L}, with as many entries as
columns in the model matrix \code{Z}. The values of \code{L} set to \code{T}
are fixed at \eqn{\beta_{10}} under the null. The values of \code{L} set to
\code{F} are estimated. The weight vector \code{w} should contain one entry for each
element of \code{L} set to \code{T}.
}
\examples{
y = D$y;
Z = D$Z;
# Overall test for covariate effects, should reject.
wScore.nlm(y=y,Z=Z,L=c(FALSE,TRUE,TRUE,TRUE))$p;
# Overall test, directed against b3
wScore.nlm(y=y,Z=Z,L=c(FALSE,TRUE,TRUE,TRUE),w=c(1/5,1/5,3/5))$p;
# Joint test of b1=0 and b2=0, directed against b1
wScore.nlm(y=y,Z=Z,L=c(FALSE,TRUE,TRUE,FALSE),w=c(3/4,1/4))$p;
# Joint test of b1=0 and b3=0, directed against b1
wScore.nlm(y=y,Z=Z,L=c(FALSE,TRUE,FALSE,TRUE),w=c(3/4,1/4))$p;
}
